[ad-mapper] mapper cache sync logic improved to handle api failure [modelservice] api change to take input time as seconds [ad-manager] renamed methods

+++ anomdetect/src/main/java/com/expedia/adaptivealerting/anomdetect/DetectorSource.java
@@ -54,7 +54,7 @@ public interface DetectorSource {
      */
     List<UUID> findUpdatedDetectors(long timePeriod);
 
-    List<DetectorMapping> findUpdatedDetectorMappings(int timePeriod);
+    List<DetectorMapping> findUpdatedDetectorMappings(long timePeriod);
 
     DetectorMatchResponse findMatchingDetectorMappings(List<Map<String, String>> metricTags);
 }
